+++ strawberryfields/utils/program_functions.py
@@ -12,39 +12,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 """
-This module defines and implements several utility functions and language extensions that complement
-StrawberryFields. These include:
-
-
-* **NumPy state functions**
-
-  These functions allow the calculation of various quantum states in either the Fock
-  basis (a one-dimensional array indexed by Fock state) or the Gaussian basis (returning
-  a vector of means and covariance matrix). These state calculations are NOT done in the
-  simulators, but rather in NumPy.
-
-  These are useful for generating states for use in calculating the fidelity of simulations.
-
-
-* **Random functions**
-
-  These functions generate random numbers and matrices corresponding to various
-  quantum states and operations.
-
-* **Decorators**
-
-  The :class:`~.strawberryfields.utils.operation` decorator allows functions
-  containing quantum operations acting on a qumode to be used as an
-  operation itself within a :class:`.Program` context.
-
-* **Program functions**
-
-  These functions act on :class:`.Program` instances, returning
-  or extracting information from the quantum circuit.
+This module defines and implements several utility functions that act on
+:class:`.Program` instances, returning or extracting information from the
+quantum circuit.
 """
-import collections
 import copy
-from inspect import signature
 
 try:
     import tensorflow as tf
@@ -52,517 +24,17 @@ except ImportError:
     tf_available = False
 
 import numpy as np
-from numpy.polynomial.hermite import hermval
-import scipy as sp
-from scipy.special import factorial as fac
 
 from strawberryfields.engine import LocalEngine
 from strawberryfields.program_utils import Command
 from strawberryfields.ops import Gate, Channel, Ket
 
 __all__ = [
-    "squeezed_cov",
-    "vacuum_state",
-    "coherent_state",
-    "squeezed_state",
-    "displaced_squeezed_state",
-    "fock_state",
-    "cat_state",
-    "randnc",
-    "random_covariance",
-    "random_symplectic",
-    "random_interferometer",
-    "operation",
     "is_unitary",
     "is_channel",
     "extract_unitary",
     "extract_channel",
 ]
-
-# ------------------------------------------------------------------------
-# State functions - Fock basis and Gaussian basis                |
-# ------------------------------------------------------------------------
-
-
-def squeezed_cov(r, phi, hbar=2):
-    r"""Returns the squeezed covariance matrix of a squeezed state
-
-    Args:
-        r (complex): the squeezing magnitude
-        p (float): the squeezing phase :math:`\phi`
-        hbar (float): (default 2) the value of :math:`\hbar` in the commutation
-            relation :math:`[\x,\p]=i\hbar`
-    Returns:
-        array: the squeezed state
-    """
-    cov = np.array([[np.exp(-2 * r), 0], [0, np.exp(2 * r)]]) * hbar / 2
-
-    R = np.array([[np.cos(phi / 2), -np.sin(phi / 2)], [np.sin(phi / 2), np.cos(phi / 2)]])
-
-    return np.dot(np.dot(R, cov), R.T)
-
-
-def vacuum_state(basis="fock", fock_dim=5, hbar=2.0):
-    r""" Returns the vacuum state
-
-    Args:
-        basis (str): If 'fock', calculates the initial state
-            in the Fock basis. If 'gaussian', returns the
-            vector of means and the covariance matrix.
-        fock_dim (int): the size of the truncated Fock basis if
-            using the Fock basis representation
-        hbar (float): (default 2) the value of :math:`\hbar` in the commutation
-            relation :math:`[\x,\p]=i\hbar`
-    Returns:
-        array: the vacuum state
-    """
-    if basis == "fock":
-        state = np.zeros((fock_dim))
-        state[0] = 1.0
-
-    elif basis == "gaussian":
-        means = np.zeros((2))
-        cov = np.identity(2) * hbar / 2
-        state = [means, cov]
-
-    return state
-
-
-def coherent_state(a, basis="fock", fock_dim=5, hbar=2.0):
-    r""" Returns the coherent state
-
-    This can be returned either in the Fock basis,
-
-    .. math::
-
-        |\alpha\rangle = e^{-|\alpha|^2/2} \sum_{n=0}^\infty
-        \frac{\alpha^n}{\sqrt{n!}}|n\rangle
-
-    or as a Gaussian:
-
-    .. math::
-
-        \mu = (\text{Re}(\alpha),\text{Im}(\alpha)),~~~\sigma = I
-
-    where :math:`\alpha` is the displacement.
-
-    Args:
-        a (complex) : the displacement
-        basis (str): If 'fock', calculates the initial state
-            in the Fock basis. If 'gaussian', returns the
-            vector of means and the covariance matrix.
-        fock_dim (int): the size of the truncated Fock basis if
-            using the Fock basis representation
-        hbar (float): (default 2) the value of :math:`\hbar` in the commutation
-            relation :math:`[\x,\p]=i\hbar`
-    Returns:
-        array: the coherent state
-    """
-    if basis == "fock":
-        state = np.array(
-            [np.exp(-0.5 * np.abs(a) ** 2) * a ** n / np.sqrt(fac(n)) for n in range(fock_dim)]
-        )
-
-    elif basis == "gaussian":
-        means = np.array([a.real, a.imag]) * np.sqrt(2 * hbar)
-        cov = np.identity(2) * hbar / 2
-        state = [means, cov]
-
-    return state
-
-
-def squeezed_state(r, p, basis="fock", fock_dim=5, hbar=2.0):
-    r""" Returns the squeezed state
-
-    This can be returned either in the Fock basis,
-
-    .. math::
-
-        |z\rangle = \frac{1}{\sqrt{\cosh(r)}}\sum_{n=0}^\infty
-        \frac{\sqrt{(2n)!}}{2^n n!}(-e^{i\phi}\tanh(r))^n|2n\rangle
-
-    or as a Gaussian:
-
-    .. math:: \mu = (0,0)
-
-    .. math::
-        :nowrap:
-
-        \begin{align*}
-            \sigma = R(\phi/2)\begin{bmatrix}e^{-2r} & 0 \\0 & e^{2r} \\\end{bmatrix}R(\phi/2)^T
-        \end{align*}
-
-    where :math:`z = re^{i\phi}` is the squeezing factor.
-
-    Args:
-        r (complex): the squeezing magnitude
-        p (float): the squeezing phase :math:`\phi`
-        basis (str): If 'fock', calculates the initial state
-            in the Fock basis. If 'gaussian', returns the
-            vector of means and the covariance matrix.
-        fock_dim (int): the size of the truncated Fock basis if
-            using the Fock basis representation
-        hbar (float): (default 2) the value of :math:`\hbar` in the commutation
-            relation :math:`[\x,\p]=i\hbar`
-    Returns:
-        array: the squeezed state
-    """
-    phi = p
-
-    if basis == "fock":
-
-        def ket(n):
-            """Squeezed state kets"""
-            return (np.sqrt(fac(2 * n)) / (2 ** n * fac(n))) * (-np.exp(1j * phi) * np.tanh(r)) ** n
-
-        state = np.array([ket(n // 2) if n % 2 == 0 else 0.0 for n in range(fock_dim)])
-        state *= np.sqrt(1 / np.cosh(r))
-
-    elif basis == "gaussian":
-        means = np.zeros((2))
-        state = [means, squeezed_cov(r, phi, hbar)]
-
-    return state
-
-
-def displaced_squeezed_state(a, r, phi, basis="fock", fock_dim=5, hbar=2.0):
-    r""" Returns the squeezed coherent state
-
-    This can be returned either in the Fock basis,
-
-    .. math::
-
-        |\alpha,z\rangle = e^{-\frac{1}{2}|\alpha|^2-\frac{1}{2}{\alpha^*}^2 e^{i\phi}\tanh{(r)}}
-        \sum_{n=0}^\infty\frac{\left[\frac{1}{2}e^{i\phi}\tanh(r)\right]^{n/2}}{\sqrt{n!\cosh(r)}}
-        H_n\left[ \frac{\alpha\cosh(r)+\alpha^*e^{i\phi}\sinh(r)}{\sqrt{e^{i\phi}\sinh(2r)}} \right]|n\rangle
-
-    where :math:`H_n(x)` is the Hermite polynomial, or as a Gaussian:
-
-    .. math:: \mu = (\text{Re}(\alpha),\text{Im}(\alpha))
-
-    .. math::
-        :nowrap:
-
-        \begin{align*}
-            \sigma = R(\phi/2)\begin{bmatrix}e^{-2r} & 0 \\0 & e^{2r} \\\end{bmatrix}R(\phi/2)^T
-        \end{align*}
-
-    where :math:`z = re^{i\phi}` is the squeezing factor
-    and :math:`\alpha` is the displacement.
-
-    Args:
-        a (complex): the displacement
-        r (complex): the squeezing magnitude
-        phi (float): the squeezing phase :math:`\phi`
-        basis (str): If 'fock', calculates the initial state
-            in the Fock basis. If 'gaussian', returns the
-            vector of means and the covariance matrix.
-        fock_dim (int): the size of the truncated Fock basis if
-            using the Fock basis representation
-        hbar (float): (default 2) the value of :math:`\hbar` in the commutation
-            relation :math:`[\x,\p]=i\hbar`
-    Returns:
-        array: the squeezed coherent state
-    """
-    # pylint: disable=too-many-arguments
-    if basis == "fock":
-
-        if r != 0:
-            phase_factor = np.exp(1j * phi)
-            ch = np.cosh(r)
-            sh = np.sinh(r)
-            th = np.tanh(r)
-
-            gamma = a * ch + np.conj(a) * phase_factor * sh
-            N = np.exp(-0.5 * np.abs(a) ** 2 - 0.5 * np.conj(a) ** 2 * phase_factor * th)
-
-            coeff = np.diag(
-                [
-                    (0.5 * phase_factor * th) ** (n / 2) / np.sqrt(fac(n) * ch)
-                    for n in range(fock_dim)
-                ]
-            )
-
-            vec = [hermval(gamma / np.sqrt(phase_factor * np.sinh(2 * r)), row) for row in coeff]
-
-            state = N * np.array(vec)
-
-        else:
-            state = coherent_state(a, basis="fock", fock_dim=fock_dim)  # pragma: no cover
-
-    elif basis == "gaussian":
-        means = np.array([a.real, a.imag]) * np.sqrt(2 * hbar)
-        state = [means, squeezed_cov(r, phi, hbar)]
-
-    return state
-
-
-# ------------------------------------------------------------------------
-# State functions - Fock basis only                              |
-# ------------------------------------------------------------------------
-
-
-def fock_state(n, fock_dim=5):
-    r""" Returns the Fock state
-
-    Args:
-        n (int): the occupation number
-        fock_dim (int): the size of the truncated Fock basis
-    Returns:
-        array: the Fock state
-    """
-    ket = np.zeros((fock_dim))
-    ket[n] = 1.0
-    return ket
-
-
-def cat_state(a, p=0, fock_dim=5):
-    r""" Returns the cat state
-
-    .. math::
-
-        |cat\rangle = \frac{1}{\sqrt{2(1+e^{-2|\alpha|^2}\cos(\phi))}}
-        \left(|\alpha\rangle +e^{i\phi}|-\alpha\rangle\right)
-
-    with the even cat state given for :math:`\phi=0`, and the odd
-    cat state given for :math:`\phi=\pi`.
-
-    Args:
-        a (complex): the displacement
-        p (float): parity, where :math:`\phi=p\pi`. ``p=0`` corresponds to an even
-            cat state, and ``p=1`` an odd cat state
-        fock_dim (int): the size of the truncated Fock basis
-    Returns:
-        array: the cat state
-    """
-    # p=0 if even, p=pi if odd
-    phi = np.pi * p
-
-    # normalisation constant
-    temp = np.exp(-0.5 * np.abs(a) ** 2)
-    N = temp / np.sqrt(2 * (1 + np.cos(phi) * temp ** 4))
-
-    # coherent states
-    k = np.arange(fock_dim)
-    c1 = (a ** k) / np.sqrt(fac(k))
-    c2 = ((-a) ** k) / np.sqrt(fac(k))
-
-    # add them up with a relative phase
-    ket = (c1 + np.exp(1j * phi) * c2) * N
-
-    return ket
-
-
-# ------------------------------------------------------------------------
-# Random numbers and matrices                                           |
-# ------------------------------------------------------------------------
-
-
-def randnc(*arg):
-    """Normally distributed array of random complex numbers."""
-    return np.random.randn(*arg) + 1j * np.random.randn(*arg)
-
-
-def random_covariance(N, hbar=2, pure=False, block_diag=False):
-    r"""Random covariance matrix.
-
-    Args:
-        N (int): number of modes
-        hbar (float): the value of :math:`\hbar` to use in the definition
-            of the quadrature operators :math:`\x` and :math:`\p`
-        pure (bool): If True, a random covariance matrix corresponding
-            to a pure state is returned.
-        block_diag (bool): If True, uses passive Gaussian transformations that are orthogonal
-            instead of unitary. This implies that the positions :math:`q` do not mix with
-            the momenta :math:`p` and thus the covariance matrix is block diagonal.
-    Returns:
-        array: random :math:`2N\times 2N` covariance matrix
-    """
-    S = random_symplectic(N, block_diag=block_diag)
-
-    if pure:
-        return (hbar / 2) * S @ S.T
-
-    nbar = 2 * np.abs(np.random.random(N)) + 1
-    Vth = (hbar / 2) * np.diag(np.concatenate([nbar, nbar]))
-
-    return S @ Vth @ S.T
-
-
-def random_symplectic(N, passive=False, block_diag=False, scale=1.0):
-    r"""Random symplectic matrix representing a Gaussian transformation.
-
-    The squeezing parameters :math:`r` for active transformations are randomly
-    sampled from the standard normal distribution, while passive transformations
-    are randomly sampled from the Haar measure. Note that for the Symplectic
-    group there is no notion of Haar measure since this is group is not compact.
-
-    Args:
-        N (int): number of modes
-        passive (bool): If True, returns a passive Gaussian transformation (i.e.,
-            one that preserves photon number). If False (default), returns an active
-            transformation.
-        block_diag (bool): If True, uses passive Gaussian transformations that are orthogonal
-            instead of unitary. This implies that the positions :math:`q` do not mix with
-            the momenta :math:`p` and thus the symplectic operator is block diagonal
-        scale (float): Sets the scale of the random values used as squeezing parameters.
-            They will range from 0 to :math:`\sqrt{2}\texttt{scale}`
-
-    Returns:
-        array: random :math:`2N\times 2N` symplectic matrix
-    """
-    U = random_interferometer(N, real=block_diag)
-    O = np.vstack([np.hstack([U.real, -U.imag]), np.hstack([U.imag, U.real])])
-
-    if passive:
-        return O
-
-    U = random_interferometer(N, real=block_diag)
-    P = np.vstack([np.hstack([U.real, -U.imag]), np.hstack([U.imag, U.real])])
-
-    r = scale * np.abs(randnc(N))
-    Sq = np.diag(np.concatenate([np.exp(-r), np.exp(r)]))
-
-    return O @ Sq @ P
-
-
-def random_interferometer(N, real=False):
-    r"""Random unitary matrix representing an interferometer.
-
-    For more details, see :cite:`mezzadri2006`.
-
-    Args:
-        N (int): number of modes
-        real (bool): return a random real orthogonal matrix
-
-    Returns:
-        array: random :math:`N\times N` unitary distributed with the Haar measure
-    """
-    if real:
-        z = np.random.randn(N, N)
-    else:
-        z = randnc(N, N) / np.sqrt(2.0)
-    q, r = sp.linalg.qr(z)
-    d = sp.diagonal(r)
-    ph = d / np.abs(d)
-    U = np.multiply(q, ph, q)
-    return U
-
-
-# ------------------------------------------------------------------------
-# Decorators                                                            |
-# ------------------------------------------------------------------------
-
-
-class operation:
-    """Groups a sequence of gates into a single operation to be used
-    within a Program context.
-
-    For example:
-
-    .. code-block:: python
-
-        @sf.operation(3)
-        def custom_operation(v1, v2, q):
-            CZgate(v1) | (q[0], q[1])
-            Vgate(v2) | q[2]
-
-    Here, the ``operation`` decorator must recieve an argument
-    detailing the number of subsystems the resulting custom
-    operation acts on.
-
-    The function it acts on can contain arbitrary
-    Python and Blackbird code that may normally be placed within a
-    Program context. Note that it must always accept the register
-    ``q`` it acts on as the *last* argument of the function.
-
-    Once defined, it can be used like any other quantum operation:
-
-    .. code-block:: python
-
-        prog = sf.Program(3)
-        with prog.context as q:
-            custom_operation(0.5719, 2.0603) | (q[0], q[1], q[3])
-
-    Note that here, we do not pass the register ``q`` directly
-    to the function - instead, it is defined on the right hand side
-    of the ``|`` operation, like all other Blackbird code.
-
-    Args:
-        ns (int): number of subsystems required by the operation
-    """
-
-    def __init__(self, ns):
-        self.ns = ns
-        self.func = None
-        self.args = None
-
-    def __or__(self, reg):
-        """Apply the operation to a part of a quantum register.
-
-        Redirects the execution flow to the wrapped function.
-
-        Args:
-            reg (RegRef, Sequence[RegRef]): subsystem(s) the operation is acting on
-
-        Returns:
-            list[RegRef]: subsystem list as RegRefs
-        """
-        if (not reg) or (not self.ns):
-            raise ValueError("Wrong number of subsystems")
-
-        reg_len = 1
-        if isinstance(reg, collections.abc.Sized):
-            reg_len = len(reg)
-
-        if reg_len != self.ns:
-            raise ValueError("Wrong number of subsystems")
-
-        return self._call_function(reg)
-
-    def _call_function(self, reg):
-        """Executes the wrapped function and passes the quantum registers.
-
-        Args:
-            reg (RegRef, Sequence[RegRef]): subsystem(s) the operation is acting on
-
-        Returns:
-            list[RegRef]: subsystem list as RegRefs
-        """
-        func_sig = signature(self.func)
-        num_params = len(func_sig.parameters)
-
-        if num_params == 0:
-            raise ValueError("Operation must receive the qumode register as an argument.")
-
-        if num_params != len(self.args) + 1:
-            raise ValueError("Mismatch in the number of arguments")
-
-        # pass parameters and subsystems to the function
-        if num_params == 1:
-            self.func(reg)
-        else:
-            self.func(*self.args, reg)
-
-        return reg
-
-    def __call__(self, func):
-        self.func = func
-
-        def f_proxy(*args):
-            """
-            Proxy for function execution. Function will actually execute in __or__
-            """
-            self.args = args
-            return self
-
-        return f_proxy
-
-
-# =================================================
-# Program functions
-# =================================================
-
 
 def is_unitary(prog):
     """True iff all the operations in the program are unitary.
